hw2/results/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "filatov_latex_generator>=0.1.0",
]